backend-agent/requirements.txt
fastapi
uvicorn
pydantic
supabase
google-genai>=1.47.0
python-dotenv
httpx
